move firebase under services dir
