+++ packages/nextjs/app/api/builds/upload-image/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from "next/server";
 import mime from "mime-types";
-import { bucket } from "~~/utils/firebase";
+import { bucket } from "~~/services/firebase";
 
 // Export route segment config for API routes in App Router
 export const dynamic = "force-dynamic";
