+++ packages/nextjs/services/firebase.ts
